No reconnect on ping with inflight requests
There's an inherent race waiting for pings, and having items in the
inflight queue can be a completely normal thing. Consider that there
might be a ready_channel() call running while a consumer is sending
requests. The timeout would have been running if the client was
previously idle. Now, the timer fires. The next spin through the event
loop would fire the timer even though there is items in the queue and
even if it had been cancelled (the event was delivered to epoll either
way).